notes on fixing the 14kb/launch VRAM leak due to use of separate CUDA stream for each launch, plus change non-controllable per launch logging to be under SEvt__MINIMAL control

+++ sysrap/smonitor.py
@@ -12,7 +12,7 @@ import matplotlib.pyplot as mp
 SIZE=np.array([1280, 720])
 
 if __name__ == '__main__':
-    mon = np.load("smonitor.npy")
+    mon = np.load("smonitor.npy").astype(np.int64) 
 
     stamp = mon[:,0]
     device = mon[:,1]
